+++ src/molecules/CurrentSolarPowerGenerating/index.js
@@ -3,6 +3,7 @@ import styles from './CurrentSolarPowerGenerating.module.css';
 
 import Skeleton from '../../atoms/Skeleton';
 import Alert from '../../atoms/Alert';
+import ToolTip from '../../atoms/ToolTip';
 
 import { useQuery, gql } from '@apollo/client';
 import { FormattedNumber } from 'react-intl';
@@ -30,8 +31,11 @@ export default function CurrentSolarPowerGenerating({ updatedAt }) {
 	if (error) return <Alert severity="error">{error.message}</Alert>;
 	return (
 		<div className={styles.container}>
-			{!loading ? data?.currentSolarPowerGenerating?.received && data.currentSolarPowerGenerating.received !== null ?
-				<span><FormattedNumber value={data.currentSolarPowerGenerating.received} /> W </span> : '-' : <Skeleton width="3em" /> }
+			{!loading ?
+				data?.currentSolarPowerGenerating?.received && data.currentSolarPowerGenerating.received !== null ?
+					<ToolTip title="Actual power from the sun"><span><FormattedNumber value={data.currentSolarPowerGenerating.received} /> W </span></ToolTip> :
+					<ToolTip title="No solar power generating"><span>-</span></ToolTip> :
+				<Skeleton width="3em" /> }
 		</div>
 	);
 }