+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.42.0] - 2022-12-22
+### Added
+- Tooltips for single value widgets and settings header button.
+- Docker `latest` tag for docker deploy.
+
+### Fixed
+- Smart Meter settings page when no data.
+
 ## [1.41.0] - 2022-12-16
 ### Changed
 - Improved message or alert boxes.